fx/model/requirements.txt
pandas>=2.0
numpy>=1.24
requests>=2.31
statsmodels>=0.14
matplotlib>=3.7
python-dotenv>=1.0
pandaSDMX>=1.10
yfinance>=0.2
